backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "David Ho",email = "davidsw.ho02@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "strawberry-graphql[fastapi] (>=0.265.0,<0.266.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "uvicorn[standard] (>=0.34.1,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "openai (==0.27.10)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
[tool.poetry.group.dev.dependencies]
ruff = "^0.11.5"

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["migrations", "__pycache__"]